requirements.txt
click==8.1.3
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-dotenv==0.21.1
SQLAlchemy==1.4.46
Werkzeug==2.2.2
